Add Hired Table

+++ database/migrations/2024_06_22_054602_create_hired_table.php
@@ -13,15 +13,17 @@ class CreateHiredTable extends Migration
      */
     public function up()
     {
-        Schema::create('hired', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('candidate_id');
-            $table->unsignedBigInteger('job_id');
-            $table->timestamps();
+        if (!Schema::hasTable('hired')) {
+            Schema::create('hired', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('candidate_id');
+                $table->unsignedBigInteger('job_id');
+                $table->timestamps();
 
-            $table->foreign('candidate_id')->references('id')->on('candidates')->onDelete('cascade');
-            $table->foreign('job_id')->references('id')->on('jobs')->onDelete('cascade');
-        });
+                $table->foreign('candidate_id')->references('id')->on('candidates')->onDelete('cascade');
+                $table->foreign('job_id')->references('id')->on('jobs')->onDelete('cascade');
+            });
+        }
     }
 
     /**
